feat(collector): Add command line parser to collector
It allows passing custom log level and the download directory to the
collector.

Also, the code that initializes the Elastic Search index was moved
from the module scope to a function and the collector now does the job
of setting that up. Not sure if that should be done separately.